requirements.txt
transformers>=4.36.2
diffusers>=0.25.0
openai>=1.3.9
streamlit>=1.29.0
ctransformers==0.2.24
python-dotenv
Pillow>=9.3.0
accelerate>=0.25.0
cvzone>=1.6.1
pypdf
peft